@@ -19,8 +19,9 @@ c = Unwrap(b)  // a == c
 // Wrap with context.
 url := "http://host/..."
 d := e1.Wrap(
-    a, "Web request failed."
-    "url", url)
+
+	a, "Web request failed."
+	"url", url)
 
 d.Error()   // "Web request failed. caused by: 'No route to host'"
 d.Context() // []string{"url", "http://host/..."}
