@@ -8,7 +8,6 @@ import (
 	"strings"
 )
 
-//
 // Create a new wrapped error.
 func New(m string, kvpair ...interface{}) error {
 	return Wrap(
@@ -16,7 +15,6 @@ func New(m string, kvpair ...interface{}) error {
 		kvpair...)
 }
 
-//
 // Wrap an error.
 // Returns `err` when err is `nil` or *Error.
 func Wrap(err error, kvpair ...interface{}) error {
@@ -53,7 +51,6 @@ func Wrap(err error, kvpair ...interface{}) error {
 	return newError
 }
 
-//
 // Unwrap an error.
 // Returns: the original error when not wrapped.
 func Unwrap(err error) (out error) {
@@ -72,7 +69,6 @@ func Unwrap(err error) (out error) {
 	return
 }
 
-//
 // Error.
 // Wraps a root cause error and captures
 // the stack.
@@ -87,7 +83,6 @@ type Error struct {
 	stack []string
 }
 
-//
 // Error description.
 func (e Error) Error() string {
 	if len(e.description) > 0 {
@@ -97,31 +92,28 @@ func (e Error) Error() string {
 	}
 }
 
-//
 // Error stack trace.
 // Format:
-//   package.Function()
-//     file:line
-//   package.Function()
-//     file:line
-//   ...
+//
+//	package.Function()
+//	  file:line
+//	package.Function()
+//	  file:line
+//	...
 func (e Error) Stack() string {
 	return strings.Join(e.stack, "\n")
 }
 
-//
 // Get `context` key/value pairs.
 func (e Error) Context() []interface{} {
 	return e.context
 }
 
-//
 // Unwrap the error.
 func (e Error) Unwrap() error {
 	return Unwrap(e.wrapped)
 }
 
-//
 // Append context.
 // And odd number of context is interpreted as:
 // a description followed by an even number of key value pairs.
@@ -143,7 +135,6 @@ func (e *Error) append(kvpair []interface{}) {
 	e.context = append(e.context, kvpair...)
 }
 
-//
 // Build caused-by.
 func (e *Error) causedBy(error, caused string) string {
 	return fmt.Sprintf(
