@@ -1,7 +1,8 @@
 // Container
-//   |__Collector
-//   |__Collector
-//   |__Collector
+//
+//	|__Collector
+//	|__Collector
+//	|__Collector
 //
 // The container is a collection of data model collectors.
 // Each collector is responsible for ensuring that changes made
@@ -9,7 +10,6 @@
 // goal is for the data model to be eventually consistent.
 package container
 
-//
 // Build a new container.
 func New() *Container {
 	return &Container{
