@@ -2,6 +2,9 @@ package ocp
 
 import (
 	"context"
+	"path"
+	"time"
+
 	"github.com/go-logr/logr"
 	liberr "github.com/konveyor/controller/pkg/error"
 	libmodel "github.com/konveyor/controller/pkg/inventory/model"
@@ -11,21 +14,19 @@ import (
 	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes/scheme"
 	"k8s.io/client-go/rest"
-	"path"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/manager"
+	"sigs.k8s.io/controller-runtime/pkg/metrics/server"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 	"sigs.k8s.io/controller-runtime/pkg/source"
-	"time"
 )
 
 const (
 	RetryDelay = time.Second * 5
 )
 
-//
 // Cluster.
 type Cluster interface {
 	meta.Object
@@ -34,7 +35,6 @@ type Cluster interface {
 	RestCfg(*core.Secret) *rest.Config
 }
 
-//
 // An OpenShift collector.
 type Collector struct {
 	// The cluster CR.
@@ -66,7 +66,6 @@ type Collector struct {
 	cancel func()
 }
 
-//
 // New collector.
 func New(
 	db libmodel.DB,
@@ -74,7 +73,7 @@ func New(
 	secret *core.Secret,
 	collections ...Collection) *Collector {
 	//
-	log := logging.WithName("collector|ocp").WithValues(
+	log := logging.WithName("collector|ocp").Real.WithValues(
 		"cluster",
 		path.Join(
 			cluster.GetNamespace(),
@@ -88,43 +87,36 @@ func New(
 	}
 }
 
-//
 // The name.
 func (r *Collector) Name() string {
 	return r.cluster.GetName()
 }
 
-//
 // The owner.
 func (r *Collector) Owner() meta.Object {
 	return r.cluster
 }
 
-//
 // Get the DB.
 func (r *Collector) DB() libmodel.DB {
 	return r.db
 }
 
-//
 // Get the Client.
 func (r *Collector) Client() client.Client {
 	return r.client
 }
 
-//
 // Reset.
 func (r *Collector) Reset() {
 	r.parity = false
 }
 
-//
 // Collector has achieved parity.
 func (r *Collector) HasParity() bool {
 	return r.parity
 }
 
-//
 // Update the versionThreshold
 func (r *Collector) UpdateThreshold(m libmodel.Model) {
 	if m, cast := m.(interface{ ResourceVersion() uint64 }); cast {
@@ -140,7 +132,6 @@ func (r *Collector) Test() error {
 	return r.buildClient()
 }
 
-//
 // Start the collector.
 func (r *Collector) Start() error {
 	ctx := context.Background()
@@ -175,12 +166,11 @@ func (r *Collector) Start() error {
 	return nil
 }
 
-//
 // Start details.
-//   1. Build and start the manager.
-//   2. Reconcile all of the collections.
-//   3. Mark parity.
-//   4. Start apply events (coroutine).
+//  1. Build and start the manager.
+//  2. Reconcile all of the collections.
+//  3. Mark parity.
+//  4. Start apply events (coroutine).
 func (r *Collector) start(ctx context.Context) (err error) {
 	r.versionThreshold = 0
 	r.eventChannel = make(chan ModelEvent, 100)
@@ -200,7 +190,7 @@ func (r *Collector) start(ctx context.Context) (err error) {
 	if err != nil {
 		return
 	}
-	go r.manager.Start(r.stopChannel)
+	go r.manager.Start(ctx)
 	err = r.reconcileCollections(ctx)
 	if err != nil {
 		return
@@ -215,7 +205,6 @@ func (r *Collector) start(ctx context.Context) (err error) {
 	return
 }
 
-//
 // Reconcile collections.
 func (r *Collector) reconcileCollections(ctx context.Context) (err error) {
 	mark := time.Now()
@@ -241,11 +230,10 @@ func (r *Collector) reconcileCollections(ctx context.Context) (err error) {
 	return
 }
 
-//
 // Shutdown the collector.
-//   1. Close manager stop channel.
-//   2. Close watch event coroutine channel.
-//   3. Cancel the context.
+//  1. Close manager stop channel.
+//  2. Close watch event coroutine channel.
+//  3. Cancel the context.
 func (r *Collector) Shutdown() {
 	r.log.V(3).Info("shutdown.")
 	r.terminate()
@@ -254,7 +242,6 @@ func (r *Collector) Shutdown() {
 	}
 }
 
-//
 // Terminate coroutines.
 func (r *Collector) terminate() {
 	defer func() {
@@ -264,7 +251,6 @@ func (r *Collector) terminate() {
 	close(r.eventChannel)
 }
 
-//
 // Enqueue create model event.
 // Used by watch predicates.
 // Swallow panic: send on closed channel.
@@ -277,7 +263,6 @@ func (r *Collector) Create(m libmodel.Model) {
 	r.eventChannel <- ModelEvent{}.Create(m)
 }
 
-//
 // Enqueue update model event.
 // Used by watch predicates.
 // Swallow panic: send on closed channel.
@@ -290,7 +275,6 @@ func (r *Collector) Update(m libmodel.Model) {
 	r.eventChannel <- ModelEvent{}.Update(m)
 }
 
-//
 // Enqueue delete model event.
 // Used by watch predicates.
 // Swallow panic: send on closed channel.
@@ -303,13 +287,14 @@ func (r *Collector) Delete(m libmodel.Model) {
 	r.eventChannel <- ModelEvent{}.Delete(m)
 }
 
-//
 // Build the k8s manager.
 func (r *Collector) buildManager() (err error) {
 	r.manager, err = manager.New(
 		r.cluster.RestCfg(r.secret),
 		manager.Options{
-			MetricsBindAddress: "0",
+			Metrics: server.Options{
+				BindAddress: "0",
+			},
 		})
 	if err != nil {
 		err = liberr.Wrap(err)
@@ -327,9 +312,7 @@ func (r *Collector) buildManager() (err error) {
 	}
 	for _, collection := range r.collections {
 		err = dsController.Watch(
-			&source.Kind{
-				Type: collection.Object(),
-			},
+			source.Kind(r.manager.GetCache(), collection.Object()),
 			&handler.EnqueueRequestForObject{},
 			collection)
 		if err != nil {
@@ -341,7 +324,6 @@ func (r *Collector) buildManager() (err error) {
 	return
 }
 
-//
 // Build non-cached client.
 func (r *Collector) buildClient() (err error) {
 	r.client, err = client.New(
@@ -353,7 +335,6 @@ func (r *Collector) buildClient() (err error) {
 	return
 }
 
-//
 // Apply model events.
 func (r *Collector) applyEvents() {
 	r.log.V(3).Info("apply started.")
@@ -367,13 +348,11 @@ func (r *Collector) applyEvents() {
 	}
 }
 
-//
 // Never called.
-func (r *Collector) Reconcile(request reconcile.Request) (reconcile.Result, error) {
+func (r *Collector) Reconcile(ctx context.Context, request reconcile.Request) (reconcile.Result, error) {
 	return reconcile.Result{}, nil
 }
 
-//
 // Model event.
 // Used with `eventChannel`.
 type ModelEvent struct {
@@ -386,7 +365,6 @@ type ModelEvent struct {
 	action byte
 }
 
-//
 // Apply the change to the DB.
 func (r *ModelEvent) Apply(rl *Collector) (err error) {
 	tx, err := rl.db.Begin()
@@ -448,7 +426,6 @@ func (r *ModelEvent) Apply(rl *Collector) (err error) {
 	return
 }
 
-//
 // Set the event model and action.
 func (r ModelEvent) Create(m libmodel.Model) ModelEvent {
 	r.model = m
@@ -456,7 +433,6 @@ func (r ModelEvent) Create(m libmodel.Model) ModelEvent {
 	return r
 }
 
-//
 // Set the event model and action.
 func (r ModelEvent) Update(m libmodel.Model) ModelEvent {
 	r.model = m
@@ -464,7 +440,6 @@ func (r ModelEvent) Update(m libmodel.Model) ModelEvent {
 	return r
 }
 
-//
 // Set the event model and action.
 func (r ModelEvent) Delete(m libmodel.Model) ModelEvent {
 	r.model = m
