@@ -2,6 +2,10 @@ package web
 
 import (
 	"fmt"
+	"net/http"
+	"strconv"
+	"time"
+
 	"github.com/gin-gonic/gin"
 	"github.com/go-logr/logr"
 	"github.com/gorilla/websocket"
@@ -10,26 +14,20 @@ import (
 	"github.com/konveyor/controller/pkg/inventory/model"
 	"github.com/konveyor/controller/pkg/logging"
 	"github.com/konveyor/controller/pkg/ref"
-	"net/http"
-	"strconv"
-	"time"
 )
 
-//
 // Web request handler.
 type RequestHandler interface {
 	// Add routes to the `gin` router.
 	AddRoutes(*gin.Engine)
 }
 
-//
 // Paged handler.
 type Paged struct {
 	// The `page` parameter passed in the request.
 	Page model.Page
 }
 
-//
 // Prepare the handler to fulfil the request.
 // Set the `page` field using passed parameters.
 func (h *Paged) Prepare(ctx *gin.Context) int {
@@ -41,7 +39,6 @@ func (h *Paged) Prepare(ctx *gin.Context) int {
 	return http.StatusOK
 }
 
-//
 // Set the `page` field.
 func (h *Paged) setPage(ctx *gin.Context) int {
 	q := ctx.Request.URL.Query()
@@ -70,12 +67,10 @@ func (h *Paged) setPage(ctx *gin.Context) int {
 	return http.StatusOK
 }
 
-//
 // Parity (not-partial) request handler.
 type Parity struct {
 }
 
-//
 // Ensure collector has achieved parity.
 func (c *Parity) EnsureParity(r container.Collector, w time.Duration) int {
 	wait := w
@@ -96,11 +91,9 @@ func (c *Parity) EnsureParity(r container.Collector, w time.Duration) int {
 	return http.StatusPartialContent
 }
 
-//
 // Watched resource builder.
 type ResourceBuilder func(model.Model) interface{}
 
-//
 // Event
 type Event struct {
 	// ID
@@ -115,7 +108,6 @@ type Event struct {
 	Updated interface{}
 }
 
-//
 // String representation.
 func (r *Event) String() string {
 	action := "unknown"
@@ -146,7 +138,6 @@ func (r *Event) String() string {
 		kind)
 }
 
-//
 // Watch (event) writer.
 // The writer is model event handler. Each event
 // is send (forwarded) to the watch client.  This
@@ -164,13 +155,11 @@ type WatchWriter struct {
 	done bool
 }
 
-//
 // Watch options.
 func (r *WatchWriter) Options() model.WatchOptions {
 	return r.options
 }
 
-//
 // Start the writer.
 // Detect connection closed by peer or broken
 // and end the watch.
@@ -201,7 +190,6 @@ func (r *WatchWriter) Start(watch *model.Watch) {
 	}()
 }
 
-//
 // Watch has started.
 func (r *WatchWriter) Started(watchID uint64) {
 	r.log.V(3).Info("event: started.")
@@ -211,7 +199,6 @@ func (r *WatchWriter) Started(watchID uint64) {
 	})
 }
 
-//
 // Watch has parity.
 func (r *WatchWriter) Parity() {
 	r.log.V(3).Info("event: parity.")
@@ -220,7 +207,6 @@ func (r *WatchWriter) Parity() {
 	})
 }
 
-//
 // A model has been created.
 func (r *WatchWriter) Created(event model.Event) {
 	r.log.V(5).Info(
@@ -230,7 +216,6 @@ func (r *WatchWriter) Created(event model.Event) {
 	r.send(event)
 }
 
-//
 // A model has been updated.
 func (r *WatchWriter) Updated(event model.Event) {
 	r.log.V(5).Info(
@@ -240,7 +225,6 @@ func (r *WatchWriter) Updated(event model.Event) {
 	r.send(event)
 }
 
-//
 // A model has been deleted.
 func (r *WatchWriter) Deleted(event model.Event) {
 	r.log.V(5).Info(
@@ -250,7 +234,6 @@ func (r *WatchWriter) Deleted(event model.Event) {
 	r.send(event)
 }
 
-//
 // An error has occurred delivering an event.
 func (r *WatchWriter) Error(err error) {
 	r.log.V(3).Info(
@@ -262,7 +245,6 @@ func (r *WatchWriter) Error(err error) {
 	})
 }
 
-//
 // An event watch has ended.
 func (r *WatchWriter) End() {
 	r.log.V(3).Info("event: ended.")
@@ -274,7 +256,6 @@ func (r *WatchWriter) End() {
 	_ = r.webSocket.Close()
 }
 
-//
 // Write event to the socket.
 func (r *WatchWriter) send(e model.Event) {
 	if r.done {
@@ -302,7 +283,6 @@ func (r *WatchWriter) send(e model.Event) {
 		event)
 }
 
-//
 // Watched (handler).
 type Watched struct {
 	// Watch requested.
@@ -311,7 +291,6 @@ type Watched struct {
 	options model.WatchOptions
 }
 
-//
 // Prepare the handler to fulfil the request.
 // Set the `WatchRequest` and `snapshot` fields based on passed headers.
 // The header value is a list of options.
@@ -328,7 +307,6 @@ func (h *Watched) Prepare(ctx *gin.Context) int {
 	return http.StatusOK
 }
 
-//
 // Watch model.
 func (r *Watched) Watch(
 	ctx *gin.Context,
@@ -354,7 +332,7 @@ func (r *Watched) Watch(
 		options:   r.options,
 		webSocket: socket,
 		builder:   rb,
-		log: logging.WithName(name).WithValues(
+		log: logging.WithName(name).Real.WithValues(
 			"peer",
 			socket.RemoteAddr()),
 	}
@@ -363,7 +341,7 @@ func (r *Watched) Watch(
 		_ = socket.Close()
 		return
 	}
-	writer.log = logging.WithName(name).WithValues(
+	writer.log = logging.WithName(name).Real.WithValues(
 		"peer",
 		socket.RemoteAddr(),
 		"watch",
@@ -371,7 +349,7 @@ func (r *Watched) Watch(
 
 	writer.Start(watch)
 
-	log.V(3).Info(
+	log.Real.V(3).Info(
 		"handler: watch created.",
 		"url",
 		ctx.Request.URL,
@@ -381,7 +359,6 @@ func (r *Watched) Watch(
 	return
 }
 
-//
 // Schema (route) handler.
 type SchemaHandler struct {
 	// The `gin` router.
@@ -392,14 +369,12 @@ type SchemaHandler struct {
 	Release int
 }
 
-//
 // Add routes.
 func (h *SchemaHandler) AddRoutes(r *gin.Engine) {
 	r.GET("/schema", h.List)
 	h.router = r
 }
 
-//
 // List schema.
 func (h *SchemaHandler) List(ctx *gin.Context) {
 	type Schema struct {
@@ -419,7 +394,6 @@ func (h *SchemaHandler) List(ctx *gin.Context) {
 	ctx.JSON(http.StatusOK, schema)
 }
 
-//
 // Not supported.
 func (h SchemaHandler) Get(ctx *gin.Context) {
 	ctx.Status(http.StatusMethodNotAllowed)
