@@ -2,19 +2,18 @@ package web
 
 import (
 	"fmt"
+	"regexp"
+	"time"
+
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
 	"github.com/konveyor/controller/pkg/inventory/container"
 	"github.com/konveyor/controller/pkg/logging"
-	"regexp"
-	"time"
 )
 
-//
 // Package logger.
 var log = logging.WithName("web")
 
-//
 // Web server
 type WebServer struct {
 	// The optional port.  Default: 8080
@@ -38,7 +37,6 @@ type WebServer struct {
 	}
 }
 
-//
 // Start the web-server.
 // Initializes `gin` with routes and CORS origins.
 // Creates an http server to handle TLS
@@ -62,13 +60,12 @@ func (w *WebServer) Start(middleware ...gin.HandlerFunc) {
 		go router.Run(w.address())
 	}
 
-	log.V(3).Info(
+	log.Real.V(3).Info(
 		"web: engine started.",
 		"address",
 		w.address())
 }
 
-//
 // Determine the address.
 func (w *WebServer) address() string {
 	if w.Port == 0 {
@@ -82,7 +79,6 @@ func (w *WebServer) address() string {
 	return fmt.Sprintf(":%d", w.Port)
 }
 
-//
 // Build a REGEX for each CORS origin.
 func (w *WebServer) buildOrigins() {
 	w.allowedOrigins = []*regexp.Regexp{}
@@ -95,7 +91,6 @@ func (w *WebServer) buildOrigins() {
 	}
 }
 
-//
 // Add the routes.
 func (w *WebServer) addRoutes(r *gin.Engine) {
 	for _, h := range w.Handlers {
@@ -103,7 +98,6 @@ func (w *WebServer) addRoutes(r *gin.Engine) {
 	}
 }
 
-//
 // Called by `gin` to perform CORS authorization.
 func (w *WebServer) allow(origin string) bool {
 	for _, expr := range w.allowedOrigins {
