@@ -1,56 +1,56 @@
 package ref
 
 import (
+	"context"
+	"reflect"
+	"strings"
+
 	"github.com/konveyor/controller/pkg/logging"
 	"k8s.io/apimachinery/pkg/types"
-	"reflect"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
-	"strings"
 )
 
-//
 // Build an event handler.
 // Example:
-//   err = cnt.Watch(
-//      &source.Kind{
-//         Type: &api.Referenced{},
-//      },
-//      libref.Handler(&api.Owner{}))
+//
+//	err = cnt.Watch(
+//	   &source.Kind{
+//	      Type: &api.Referenced{},
+//	   },
+//	   libref.Handler(&api.Owner{}))
 func Handler(owner interface{}) handler.EventHandler {
 	log := logging.WithName("ref|handler")
 	ownerKind := ToKind(owner)
-	return &handler.EnqueueRequestsFromMapFunc{
-		ToRequests: handler.ToRequestsFunc(
-			func(a handler.MapObject) []reconcile.Request {
-				refKind := ToKind(a.Object)
-				list := GetRequests(ownerKind, a)
-				if len(list) > 0 {
-					log.V(4).Info(
-						"handler: request list.",
-						"referenced",
-						refKind,
-						"owner",
-						ownerKind,
-						"list",
-						list)
-				}
-				return list
-			}),
-	}
+	return handler.EnqueueRequestsFromMapFunc(
+		func(cxt context.Context, obj client.Object) []reconcile.Request {
+			list := GetRequests(obj, ownerKind)
+			if len(list) > 0 {
+				log.V(4).Info(
+					"handler: request list.",
+					"referenced",
+					obj.GetObjectKind().GroupVersionKind().Kind,
+					"owner",
+					ownerKind,
+					"list",
+					list)
+			}
+			return list
+		},
+	)
 }
 
-//
 // Impl the handler interface.
-func GetRequests(kind string, a handler.MapObject) []reconcile.Request {
+func GetRequests(obj client.Object, ownerKind string) []reconcile.Request {
 	target := Target{
-		Kind:      ToKind(a.Object),
-		Name:      a.Meta.GetName(),
-		Namespace: a.Meta.GetNamespace(),
+		Kind:      obj.GetObjectKind().GroupVersionKind().Kind,
+		Name:      obj.GetName(),
+		Namespace: obj.GetNamespace(),
 	}
 	list := []reconcile.Request{}
 	for _, owner := range Map.Find(target) {
-		if owner.Kind != kind {
+		if owner.Kind != ownerKind {
 			continue
 		}
 		list = append(
@@ -66,7 +66,6 @@ func GetRequests(kind string, a handler.MapObject) []reconcile.Request {
 	return list
 }
 
-//
 // Determine the resource Kind.
 func ToKind(resource interface{}) string {
 	t := reflect.TypeOf(resource).String()
