@@ -1,8 +1,9 @@
 package ref
 
 import (
-	"k8s.io/api/core/v1"
 	"reflect"
+
+	"k8s.io/api/core/v1"
 	"sigs.k8s.io/controller-runtime/pkg/event"
 )
 
@@ -10,26 +11,25 @@ const (
 	Tag = "ref"
 )
 
-//
 // Predicate Event Mapper
 // All ObjectReference fields with the `ref` tag will be mapped.
 //
 // Example (CRD):
-//     type Resource struct {
-//         ThingRef *v1.ObjectReference `json:"thingRef" ref:"Thing"`
-//     }
+//
+//	type Resource struct {
+//	    ThingRef *v1.ObjectReference `json:"thingRef" ref:"Thing"`
+//	}
 //
 // Example (usage):
-//     func (p Predicate) Create(e event.CreateEvent) bool {
-//         ...
-//         ref.Mapper.Create(e)
-//     }}
 //
+//	func (p Predicate) Create(e event.CreateEvent) bool {
+//	    ...
+//	    ref.Mapper.Create(e)
+//	}}
 type EventMapper struct {
 	Map *RefMap
 }
 
-//
 // Create event.
 func (r *EventMapper) Create(event event.CreateEvent) {
 	kind := ToKind(event.Object)
@@ -39,15 +39,14 @@ func (r *EventMapper) Create(event event.CreateEvent) {
 		kind)
 	refOwner := Owner{
 		Kind:      kind,
-		Namespace: event.Meta.GetNamespace(),
-		Name:      event.Meta.GetName(),
+		Namespace: event.Object.GetNamespace(),
+		Name:      event.Object.GetName(),
 	}
 	for _, ref := range r.findRefs(event.Object) {
 		r.Map.Add(refOwner, ref)
 	}
 }
 
-//
 // Update event.
 func (r *EventMapper) Update(event event.UpdateEvent) {
 	kind := ToKind(event.ObjectNew)
@@ -57,20 +56,19 @@ func (r *EventMapper) Update(event event.UpdateEvent) {
 		kind)
 	r.Map.DeleteOwner(Owner{
 		Kind:      kind,
-		Namespace: event.MetaOld.GetNamespace(),
-		Name:      event.MetaOld.GetName(),
+		Namespace: event.ObjectOld.GetNamespace(),
+		Name:      event.ObjectOld.GetName(),
 	})
 	refOwner := Owner{
 		Kind:      kind,
-		Namespace: event.MetaNew.GetNamespace(),
-		Name:      event.MetaNew.GetName(),
+		Namespace: event.ObjectNew.GetNamespace(),
+		Name:      event.ObjectNew.GetName(),
 	}
 	for _, ref := range r.findRefs(event.ObjectNew) {
 		r.Map.Add(refOwner, ref)
 	}
 }
 
-//
 // Delete Mapper.
 func (r *EventMapper) Delete(event event.DeleteEvent) {
 	kind := ToKind(event.Object)
@@ -80,12 +78,11 @@ func (r *EventMapper) Delete(event event.DeleteEvent) {
 		kind)
 	r.Map.DeleteOwner(Owner{
 		Kind:      kind,
-		Namespace: event.Meta.GetNamespace(),
-		Name:      event.Meta.GetName(),
+		Namespace: event.Object.GetNamespace(),
+		Name:      event.Object.GetName(),
 	})
 }
 
-//
 // Inspect the object for references.
 func (r *EventMapper) findRefs(object interface{}) []Target {
 	list := []Target{}
